adding remastered four lab
