+++ expected.csv
@@ -1,9 +1,9 @@
-model,resolution,connectedDevice,lampHours,descriptionOfTheInformationOutputMode,theMaximumPermissibleLampOperatingTime
-Panasonic,1920x1080,HDMI,5.0,sport,5.0
-Epson,4k,HDMI,5.0,normal,0.0
 model,resolution,connectedDevice,yearOfASale,screenSizeInInches,versionOfSmartTV,guarantee,WORKING_HOURS_PER_YEAR
 Samsung,4k,HDMI,2023.0,57.0,last,5.0,3650.0
 Samsung,8k,HDMI,2021.0,77.0,last,0.0,3650.0
+model,resolution,connectedDevice,lampHours,descriptionOfTheInformationOutputMode,theMaximumPermissibleLampOperatingTime
+Panasonic,1920x1080,HDMI,5.0,sport,5.0
+Epson,4k,HDMI,5.0,normal,0.0
 model,resolution,connectedDevice,colorOfSurface,isHasMagniteSurface,thePossibilityOfRecordingTheScreen,guarantee,WORKING_HOURS_PER_YEAR
 Prestigio,1920x1080,HDMI,black,true,false,5.0,3650.0
 Prestigio,1920x1080,HDMI,black,false,false,5.0,3650.0
