+++ src/main/java/ua/lviv/iot/algo/part1/modules/MultiBoard.java
@@ -1,4 +1,4 @@
-package ua.lviv.iot.algo.part1.lab1;
+package ua.lviv.iot.algo.part1.modules;
 
 import lombok.Getter;
 import lombok.Setter;
