+++ src/test/java/ua/lviv/iot/algo/part1/managers/ProjectorManagerTest.java
@@ -1,8 +1,13 @@
-package ua.lviv.iot.algo.part1.lab1;
+package ua.lviv.iot.algo.part1.managers;
 
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import ua.lviv.iot.algo.part1.modules.HomeTheater;
+import ua.lviv.iot.algo.part1.modules.LampProjector;
+import ua.lviv.iot.algo.part1.modules.MultiBoard;
+import ua.lviv.iot.algo.part1.modules.TreeDProjector;
 
 class ProjectorManagerTest {
 
@@ -44,6 +49,10 @@ class ProjectorManagerTest {
                         .connectedDevice("HDMI")
                         .energyConsumption(30).constructor());
     }
+    @AfterEach
+    public void tearDown() {
+        projectorManager.addProjector().clear();
+    }
 
     @Test
     void findProjectorByModel() {
@@ -60,7 +69,7 @@ class ProjectorManagerTest {
     @Test
     void findProjectorByResolution() {
         var results = projectorManager.findProjectorByResolution("1920x1080");
-        Assertions.assertEquals(2, results);
+        Assertions.assertEquals(3, results.size());
 
         for (var projector : results) {
             Assertions.assertEquals("1920x1080", projector.getResolution());
