+++ src/test/java/ua/lviv/iot/algo/part1/modules/LampProjectorTest.java
@@ -1,10 +1,7 @@
-package ua.lviv.iot.algo.part1.lab1;
+package ua.lviv.iot.algo.part1.modules;
 
-import com.jparams.verifier.tostring.ToStringVerifier;
 import org.junit.jupiter.api.Test;
-
 import java.util.ArrayList;
-
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
